requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2016.6
pytz-deprecation-shim==0.1.0.post0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
whitenoise==5.3.0
